requirement.txt
fastapi
uvicorn
pypdf
tiktoken
faiss-cpu
numpy
reportlab
openai
python-dotenv
sentence-transformers
requests
